+++ backend/src/dwyeapi/__init__.py
@@ -33,4 +33,4 @@ __all__ = [
     "is_prod",
     "logger",
 ]
-__version__ = "0.10.0"
+__version__ = "0.10.1"